--- src/domain/object/door_state.rs
@@ -1,9 +1,0 @@
-use chrono::{DateTime, Local};
-use serde_derive::{Deserialize, Serialize};
-
-#[derive(Serialize, Deserialize)]
-pub struct DoorState {
-    pub open: bool,
-    pub timestamp: DateTime<Local>,
-    pub device_id: String,
-}
